requirements.txt
numpy == 1.21.3
faiss == 1.5.3
coloredlogs == 15.0.1
osmnx == 1.1.2
networkx == 2.6.3
shapely == 1.8.0
matplotlib == 3.4.3
pyyaml == 6.0
pandas == 1.3.4
coord-convert == 0.2.1
pytorch == 1.10.1